9.14 Добавил роут для добавления пользователя менеджером
